+++ verifier-service/src/test/java/ch/admin/bj/swiyu/verifier/service/oid4vp/service/PresentationVerificationUsecaseTest.java
@@ -2,8 +2,6 @@ package ch.admin.bj.swiyu.verifier.service.oid4vp.service;
 
 import ch.admin.bj.swiyu.verifier.common.config.ApplicationProperties;
 import ch.admin.bj.swiyu.verifier.common.exception.ProcessClosedException;
-import ch.admin.bj.swiyu.verifier.common.exception.VerificationError;
-import ch.admin.bj.swiyu.verifier.common.exception.VerificationErrorResponseCode;
 import ch.admin.bj.swiyu.verifier.common.exception.VerificationException;
 import ch.admin.bj.swiyu.verifier.domain.SdJwt;
 import ch.admin.bj.swiyu.verifier.domain.management.Management;
@@ -16,6 +14,7 @@ import ch.admin.bj.swiyu.verifier.domain.management.dcql.DcqlQuery;
 import ch.admin.bj.swiyu.verifier.dto.VerificationPresentationDCQLRequestDto;
 import ch.admin.bj.swiyu.verifier.dto.VerificationPresentationRejectionDto;
 import ch.admin.bj.swiyu.verifier.service.callback.CallbackEventProducer;
+import ch.admin.bj.swiyu.verifier.service.management.ManagementMapper;
 import ch.admin.bj.swiyu.verifier.service.management.ManagementService;
 import ch.admin.bj.swiyu.verifier.service.management.ManagementTransactionalService;
 import ch.admin.bj.swiyu.verifier.service.oid4vp.DcqlPresentationVerificationService;
@@ -26,8 +25,6 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.SignedJWT;
-import jakarta.validation.Validation;
-import jakarta.validation.Validator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -46,7 +43,6 @@ import java.util.concurrent.Future;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static ch.admin.bj.swiyu.verifier.service.oid4vp.test.mock.SDJWTCredentialMock.getSDClaims;
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
@@ -73,8 +69,6 @@ class PresentationVerificationUsecaseTest {
         callbackEventProducer = mock(CallbackEventProducer.class);
         presentationVerifier = mock(PresentationVerifier.class);
         dcqlPresentationVerificationService = mock(DcqlPresentationVerificationService.class);
-
-        Validator validator = Validation.buildDefaultValidatorFactory().getValidator();
 
         presentationVerificationUsecase = new PresentationVerificationUsecase(
                 callbackEventProducer,
@@ -124,7 +118,7 @@ class PresentationVerificationUsecaseTest {
 
         presentationVerificationUsecase.receiveVerificationPresentationClientRejection(managementId, request);
 
-        verify(managementEntity).verificationFailedDueToClientRejection("User cancelled");
+        verify(managementEntity).verificationFailedDueToClientRejection(request.getErrorDescription(), ManagementMapper.toVerificationErrorResponseCode(request.getError()));
         verify(callbackEventProducer).produceEvent(managementId);
         verify(managementEntity, never()).verificationSucceeded(any());
     }
@@ -174,7 +168,7 @@ class PresentationVerificationUsecaseTest {
 
         presentationVerificationUsecase.receiveVerificationPresentationClientRejection(managementId, rejectionRequest);
 
-        verify(managementEntity).verificationFailedDueToClientRejection(errorDescription);
+        verify(managementEntity).verificationFailedDueToClientRejection(rejectionRequest.getErrorDescription(), ManagementMapper.toVerificationErrorResponseCode(rejectionRequest.getError()));
         verify(callbackEventProducer).produceEvent(managementId);
         verify(managementEntity, never()).verificationSucceeded(any());
     }
@@ -187,11 +181,11 @@ class PresentationVerificationUsecaseTest {
         VerificationPresentationRejectionDto rejectionRequest = mock(VerificationPresentationRejectionDto.class);
         when(rejectionRequest.getErrorDescription()).thenReturn("User cancelled");
 
-        var err = assertThrows(ProcessClosedException.class, () ->
+        assertThrows(ProcessClosedException.class, () ->
                 presentationVerificationUsecase.receiveVerificationPresentationClientRejection(managementId, rejectionRequest));
 
         verify(callbackEventProducer).produceEvent(managementId);
-        verify(managementEntity, never()).verificationFailedDueToClientRejection(any());
+        verify(managementEntity, never()).verificationFailedDueToClientRejection(any(), any());
     }
 
     @Test
@@ -202,11 +196,11 @@ class PresentationVerificationUsecaseTest {
         VerificationPresentationRejectionDto rejectionRequest = mock(VerificationPresentationRejectionDto.class);
         when(rejectionRequest.getErrorDescription()).thenReturn("User cancelled");
 
-        var err = assertThrows(ProcessClosedException.class, () ->
+        assertThrows(ProcessClosedException.class, () ->
                 presentationVerificationUsecase.receiveVerificationPresentationClientRejection(managementId, rejectionRequest));
 
         verify(callbackEventProducer).produceEvent(managementId);
-        verify(managementEntity, never()).verificationFailedDueToClientRejection(any());
+        verify(managementEntity, never()).verificationFailedDueToClientRejection(any(), any());
     }
 
     /**